Add flags for modifying files
This allows working around mismatched symlinks by replacing them
with the desired ones. However, this is a destructive behavior (the
previous symlink will be lost), and therefore it must be optional.

The same now applies to creating parent directories for unparented
symlinks.  Since it creates directories in the user's filesystem,
it is also somewhat destructive, and must be also optional.